improved script, readme, and organization
