+++ data/dna.dictionary.txt
@@ -1,0 +1,5 @@
+A
+T
+C
+G
+N
