+++ data/protein.dictionary.txt
@@ -1,0 +1,21 @@
+A
+C
+D
+E
+F
+G
+H
+I
+K
+L
+M
+N
+P
+Q
+R
+S
+T
+V
+W
+Y
+
